+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added methods for retrieving task lists (lending requests and requested resources).
 
+## [0.7.1] - 2018-10-23
+
+### Fixed
+
+- Fixed infinite loop in `Users::search()` when search result only contained contacts.
+
 ## [0.7.0] - 2018-09-02
 
 ### Added
@@ -65,6 +71,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 tbd.
 
-[Unreleased]: https://github.com/scriptotek/php-marc/compare/v0.6.1...HEAD
+[Unreleased]: https://github.com/scriptotek/php-marc/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/scriptotek/php-marc/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/scriptotek/php-marc/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/scriptotek/php-marc/compare/v0.6.0...v0.6.1